+++ cmd/psqldef/psqldef_test.go
@@ -34,6 +34,12 @@ func TestApply(t *testing.T) {
 
 	for name, test := range tests {
 		t.Run(name, func(t *testing.T) {
+			// This is implemented in the psqldef command layer, so it's needed for TestApply
+			if _, ok := os.LookupEnv("PGSSLMODE"); !ok {
+				os.Setenv("PGSSLMODE", "disable")
+				defer os.Unsetenv("PGSSLMODE")
+			}
+
 			// Initialize the database with test.Current
 			testutils.MustExecute("psql", "-Upostgres", "-h", "127.0.0.1", "-c", "DROP DATABASE IF EXISTS psqldef_test;")
 			testutils.MustExecute("psql", "-Upostgres", "-h", "127.0.0.1", "-c", "CREATE DATABASE psqldef_test;")
@@ -1477,9 +1483,6 @@ func TestMain(m *testing.M) {
 	if _, ok := os.LookupEnv("PGHOST"); !ok {
 		os.Setenv("PGHOST", "127.0.0.1")
 	}
-	if _, ok := os.LookupEnv("PGSSLMODE"); !ok {
-		os.Setenv("PGSSLMODE", "disable")
-	}
 
 	resetTestDatabase()
 	mustExecute("go", "build")
@@ -1577,11 +1580,6 @@ var publicAndNonPublicSchemaTestCases = []struct {
 }
 
 func connectDatabase() (adapter.Database, error) {
-	err := os.Setenv("PGSSLMODE", "disable")
-	if err != nil {
-		return nil, err
-	}
-
 	return postgres.NewDatabase(adapter.Config{
 		User:   "postgres",
 		Host:   "127.0.0.1",